Moved logic higher
